Update device-restrictions-ios.md
Adding a note that has caused some confusion. When this restriction is 'required' but there is no wifi profile coming down to the device, the device can't connect to any wifi networks. This is sometimes an intended behavior however so that the device only connects to secure.trusted networks. Important to point out for when people have issues witht he device not connecting to any wifi.