Update kernel-extensions-overview-macos.md
Adding in that a team identifier is 10 alphanumeric characters long, not more and not less. Like so - EQHXZ8M8AV or ABCDE12345. This is important to add because we might not have this validation in time for when fullscreen launces and there could be wrong formatting that leads to errors.