Add config file changes required for GCC High
The purpose of this update is to make public the required changes that GCC High tenants must make to successfully use the Intune certificate connector in the GCC High environment. The source material is from the following internal article by waluja.
https://internal.support.services.microsoft.com/en-us/help/4524460

The note section that I added just after step 5 of the Install the Intune Certificate Connector section is the note to be added.
To my knowledge, this required change for GCC High tenants is not publicly documented elsewhere.