Update device-enrollment-program-enroll-ios.md
Please add a note that explains that the end users won't be able to open the 'Learn more' link when they click it during enrollment while the device is in single app mode. When enrolling DEP devices in single app mode their end-users see the Learn more link and are prompted with “Cannot open this link” message when attempting to launch. Having a 'note' in the documentation would be helpful so admins can tell their end users that clicking the 'learn more' link results in error until they enroll.